feat(mcp): Stage 137 — add graph and blocker preview wrappers
Move acceptance-graph-derived + recurring-blocker-detection (and the 3
optional-input helpers benchmark-handoff / decision-outcome-link /
evolution-action-preview, needed by the moved graph test) into
@conclave-ai/workspace-preview with subpath exports; dashboard files
become thin re-export wrappers (zero call-site churn). Moved their tests
too.

MCP Basic snapshot wrappers: previewAcceptanceGraphSummary +
previewRecurringBlockers — pure, read-only, return {ok,kind,preview} +
boundary (mutatesState/usesHostedExecution/requiresPayment false,
derivedPreviewOnly true); recurring-blockers derives the graph internally
when not provided; malformed input does not throw. No server runtime
wiring, no network/env/mutation/LLM/payment/agent/PR-post.

workspace-preview 145/145, mcp-workspace 36/36, dashboard 248/248
(5 helper suites moved to package), monorepo typecheck 57/57, build green.
