+++ packages/mcp-workspace/src/mcp-basic-preview-tools.d.mts
@@ -17,7 +17,13 @@ export type McpPreviewError = McpPreviewBoundary & {
 
 export type McpPreviewResult = McpPreviewBoundary & {
   ok: true;
-  kind: "acceptance_map" | "stage_plan" | "agent_run_plan" | "evidence_plan";
+  kind:
+    | "acceptance_map"
+    | "stage_plan"
+    | "agent_run_plan"
+    | "evidence_plan"
+    | "acceptance_graph_summary"
+    | "recurring_blockers";
   preview: unknown;
 };
 
@@ -27,5 +33,21 @@ export function previewAcceptanceMap(input: McpPreviewInput): McpPreviewResponse
 export function previewStagePlan(input: McpPreviewInput): McpPreviewResponse;
 export function previewAgentRunPlan(input: McpPreviewInput): McpPreviewResponse;
 export function previewEvidencePlan(input: McpPreviewInput): McpPreviewResponse;
+
+export type McpGraphPreviewInput = {
+  workflowRecordId?: string;
+  title?: string;
+  sourceSummary?: string;
+  acceptanceGraphView?: unknown;
+  acceptanceMap?: unknown;
+  stagePlan?: unknown;
+  agentRunPlan?: unknown;
+  evidencePlan?: unknown;
+  decisionOutcomePreview?: unknown;
+  evolutionActionPreview?: unknown;
+};
+
+export function previewAcceptanceGraphSummary(input: McpGraphPreviewInput): McpPreviewResult;
+export function previewRecurringBlockers(input: McpGraphPreviewInput): McpPreviewResult;
 
 export const MCP_BASIC_PREVIEW_BOUNDARY: McpPreviewBoundary;
