+++ packages/workspace-preview/src/index.d.mts
@@ -14,3 +14,8 @@ export * from "./intake-acceptance-map.d.mts";
 export * from "./intake-stage-plan.d.mts";
 export * from "./intake-agent-run-plan.d.mts";
 export * from "./intake-evidence-plan.d.mts";
+export * from "./acceptance-graph-derived.d.mts";
+export * from "./recurring-blocker-detection.d.mts";
+export * from "./intake-benchmark-handoff.d.mts";
+export * from "./intake-decision-outcome-link.d.mts";
+export * from "./intake-evolution-action-preview.d.mts";
