+++ packages/workspace-preview/test/acceptance-graph-derived.test.mjs
@@ -2,13 +2,13 @@
 // from snapshots only (no graph DB, no training).
 import { test } from "node:test";
 import assert from "node:assert/strict";
-import { buildAcceptanceGraphDerivedView } from "../src/lib/acceptance-graph-derived.mjs";
-import { buildAgentRunPlan } from "../src/lib/intake-agent-run-plan.mjs";
-import { buildIntakeEvidencePlan } from "../src/lib/intake-evidence-plan.mjs";
-import { buildIntakeAcceptanceMap } from "../src/lib/intake-acceptance-map.mjs";
-import { buildIntakeStagePlan } from "../src/lib/intake-stage-plan.mjs";
-import { buildDecisionOutcomeLinkPreview } from "../src/lib/intake-decision-outcome-link.mjs";
-import { buildEvolutionActionPackPreview } from "../src/lib/intake-evolution-action-preview.mjs";
+import { buildAcceptanceGraphDerivedView } from "../src/acceptance-graph-derived.mjs";
+import { buildAgentRunPlan } from "../src/intake-agent-run-plan.mjs";
+import { buildIntakeEvidencePlan } from "../src/intake-evidence-plan.mjs";
+import { buildIntakeAcceptanceMap } from "../src/intake-acceptance-map.mjs";
+import { buildIntakeStagePlan } from "../src/intake-stage-plan.mjs";
+import { buildDecisionOutcomeLinkPreview } from "../src/intake-decision-outcome-link.mjs";
+import { buildEvolutionActionPackPreview } from "../src/intake-evolution-action-preview.mjs";
 
 const NODE_TYPES = [
   "intake", "acceptance_item", "acceptance_area", "stage", "agent_task",
