+++ packages/workspace-preview/test/intake-evolution-action-preview.test.mjs
@@ -5,13 +5,13 @@ import assert from "node:assert/strict";
 import {
   buildEvolutionActionPackPreview,
   ACTION_TYPES,
-} from "../src/lib/intake-evolution-action-preview.mjs";
-import { buildAgentRunPlan } from "../src/lib/intake-agent-run-plan.mjs";
-import { buildIntakeEvidencePlan } from "../src/lib/intake-evidence-plan.mjs";
-import { buildIntakeAcceptanceMap } from "../src/lib/intake-acceptance-map.mjs";
-import { buildIntakeStagePlan } from "../src/lib/intake-stage-plan.mjs";
-import { buildBenchmarkHandoffPreview } from "../src/lib/intake-benchmark-handoff.mjs";
-import { buildDecisionOutcomeLinkPreview } from "../src/lib/intake-decision-outcome-link.mjs";
+} from "../src/intake-evolution-action-preview.mjs";
+import { buildAgentRunPlan } from "../src/intake-agent-run-plan.mjs";
+import { buildIntakeEvidencePlan } from "../src/intake-evidence-plan.mjs";
+import { buildIntakeAcceptanceMap } from "../src/intake-acceptance-map.mjs";
+import { buildIntakeStagePlan } from "../src/intake-stage-plan.mjs";
+import { buildBenchmarkHandoffPreview } from "../src/intake-benchmark-handoff.mjs";
+import { buildDecisionOutcomeLinkPreview } from "../src/intake-decision-outcome-link.mjs";
 
 const LEVELS = ["low", "medium", "high"];
 const FORBIDDEN = ["action pack created", "fix executed", "resolved", "passed", "production ready"];
